refactor: folderize public pages
