+++ src/pages/Public/index.js
@@ -1,0 +1,1 @@
+export { default as Forbidden } from './Forbidden';
