+++ src/pages/index.js
@@ -1,4 +1,3 @@
 export * from './Protected';
 export * from './Auth';
-
-export { default as Forbidden } from './Forbidden';
+export * from './Public';
